% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{addTotalRow}
\alias{addTotalRow}
\title{Add a total row to a dataframe}
\usage{
addTotalRow(df, .na_rm = FALSE)
}
\arguments{
\item{df}{the input dataframe}

\item{.na_rm}{option to remove NA from sum}
}
\value{
Dataframe with an added row for the totals of numeric variables.
May require explicitly converting NAs.
}
\description{
Add a total row to a dataframe where the first column is the category.
}
\examples{
df <- data.frame('a' = letters, 'b' = 1:length(letters), 'c' = rep(NA, length(letters)))
addTotalRow(df)
}
