% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{asQuarterDate}
\alias{asQuarterDate}
\title{asQuarterDate}
\usage{
asQuarterDate(year, quarter, day = 1)
}
\arguments{
\item{year}{a year that makes sense}

\item{quarter}{a quarter, 1 to 4}

\item{day}{a day that makes sense}
}
\value{
a date variable according to the start date of the quarter
}
\description{
Convert discrete year and quarter to date for plotting purposes.
}
\examples{
asQuarterDate(2008,1,
}
