% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusIS_1.R
\name{bedgraph_for_IGV}
\alias{bedgraph_for_IGV}
\title{Get bedgraph file for IGV from bed file}
\usage{
bedgraph_for_IGV(bed_file, set_end_equal_start = T)
}
\arguments{
\item{bed_file}{input bed file}

\item{set_end_equal_start}{T or F; when set as T(default), the start will be set as end too.}
}
\value{
bedgraph for IGV: with 4 columns: ① chr;  ② start; ③ end; ④
}
\description{
Get bedgraph file for IGV from bed file
}
