% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{check_variables}
\alias{check_variables}
\title{Check variables in the dataset}
\usage{
check_variables(df, num_unique_vals = 3, sort_examples = FALSE)
}
\arguments{
\item{df}{the input dataframe}

\item{num_unique_vals}{number of unique values to display}

\item{sort_examples}{whether first few examples or random}
}
\value{
Dataframe with Unique Values, Percent Missing and Examples
}
\description{
Get a snapshot preview of what variables are available and what they look like.
}
\examples{
df <- data.frame('a' = letters, 'b' = 1:length(letters), 'c' = rep(NA, length(letters)))
checkVariables(df)
}
