% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadley.R
\name{commas}
\alias{commas}
\title{Collapse string with commas}
\usage{
commas(...)
}
\arguments{
\item{...}{Chr}
}
\value{
String
}
\description{
Collapse string with commas
}
\examples{
commas(letters[1:10])
}
