% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{cs}
\alias{cs}
\title{Make a character vector without quotes}
\usage{
cs(...)
}
\arguments{
\item{...}{This is coerced to a character vector}
}
\value{
A character vector of the input values.
}
\description{
Make a character vector without quotes.
Ignores if the values entered are variable names
}
\examples{
a <- TRUE
string_arb <- cs(a, b, c)
string_arb
#> "a" "b" "c"

}
