% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_dependencies.R
\name{explore_dependencies}
\alias{explore_dependencies}
\title{Explore dependencies between functions in a package.}
\usage{
explore_dependencies(pkg, .f)
}
\arguments{
\item{pkg}{Character string giving the name of a package.}

\item{.f}{Character string giving the name of a function.}
}
\value{
A plot
}
\description{
Explore dependencies between functions in a package.
}
\examples{
library(ctfs)
explore_dependencies("ctfs", "abundanceperquad")
}
\seealso{
\code{\link[DependenciesGraphs:funDependencies]{DependenciesGraphs::funDependencies()}}.
}
