% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusIS_1.R
\name{FeatureCounts_txt_to_csv}
\alias{FeatureCounts_txt_to_csv}
\title{FeatureCounts_txt_to_csv}
\usage{
FeatureCounts_txt_to_csv(
  file = "/home/SharedData/VIS/first_test/04-featureCounts_by_exon.txt",
  out_path = "/home/SharedData/VIS/first_test/05-reordered-gene-counts.csv"
)
}
\arguments{
\item{file}{input txt from Featurecounts}

\item{out_path}{output filename}
}
\value{
reordered csv file
}
\description{
process the file and reorder it.
}
