% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusIS_1.R
\name{get_logo_plot}
\alias{get_logo_plot}
\title{Title get logo plot}
\usage{
get_logo_plot(
  cleaned_mapped_fastq =
    "/home/SharedData/VIS/first_test/get-break-sub-03-trim_half_adapter.fastq"
)
}
\arguments{
\item{cleaned_mapped_fastq}{fq with the mapped reads.}
}
\value{
logo-plot
}
\description{
Title get logo plot
}
