% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_rstudio_proj.R
\name{get_my_rstudio_proj_list}
\alias{get_my_rstudio_proj_list}
\title{Get the path to the file that stores RStudio's projects list.}
\usage{
get_my_rstudio_proj_list()
}
\value{
A path.
}
\description{
Get the path to the file that stores RStudio's projects list.
}
\examples{
get_my_rstudio_proj_list()
}
