% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{\%p\%}
\alias{\%p\%}
\title{Paste magic function}
\usage{
x1 \%p\% x2
}
\arguments{
\item{x1}{string 1 to paste}

\item{x2}{string 2 to paste}
}
\description{
Paste characters.
}
\details{
Functions to help with string handling
}
