% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handy.R
\docType{package}
\name{handy-package}
\alias{handy}
\alias{handy-package}
\title{handy: Commonly Used Functions For Handling Data}
\description{
This package provides handy functions
for examining, manipulating, and outputting data.
See vignette('handy-intro') for details.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/kevinrpan/handy/issues}
}

}
\author{
\strong{Maintainer}: Kevin Pan \email{dev@kevinrpan.com}

}
\keyword{internal}
