% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handy.R
\docType{package}
\name{handy}
\alias{handy}
\alias{handy-package}
\title{handy: A package that contains several commonly used functions for cleaning and examining data.}
\description{
The handy package provides some functions that I found handy.
}
