% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{human_numbers}
\alias{human_numbers}
\alias{scale_y_human}
\alias{scale_x_human}
\alias{human_num}
\alias{human_usd}
\title{Human Numbers}
\source{
https://github.com/fdryan/R/blob/master/ggplot2_formatter.r
}
\usage{
human_numbers(x = NULL, smbl = "", signif = 1)

scale_y_human(..., smbl = "", signif = 1)

scale_x_human(..., smbl = "", signif = 1)

human_num(x)

human_usd(x)
}
\arguments{
\item{x}{a numeric vector to format,}

\item{smbl}{a symbol you'd like to prefix your numbers by e.g. "$"}

\item{signif}{the number of significant places you want the function to return}

\item{...}{other arguments passed to scale_*_continuous}
}
\value{
a character vector the same length as the input vector
}
\description{
Format numbers so they're legible for humans
Use this in ggplot for labels where you might use the comma or percent functions from the
Scales package.

Checks whether numbers are positive or negative.
Allows up to 1 significant figure
sapply used for element-wise application of the humanity function as a vector may include
numbers where billions, millions or thousands are appropriate.
}
\examples{
human_numbers(c(1000000 , 1500000, 10000000000))
human_numbers(c(1.200000e+05, -2.154660e+05, 2.387790e+05, 4.343500e+04 ,5.648675e+12), "$")
\dontrun{
ggplot2 + scale_y_continuous(labels = human_numbers)
ggplot2 + scale_x_continuous(labels = human_numbers)
ggplot2 + scale_y_human()
ggplot2 + scale_x_human()
}
}
