% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_missing_load_all.R
\name{install_missing_load_all}
\alias{install_missing_load_all}
\title{Given package names, install missing ones and load all}
\usage{
install_missing_load_all(pkg)
}
\arguments{
\item{pkg}{Package names in a character string.}
}
\value{
nothing but messages from install.packages() and library()
}
\description{
Given package names, install missing ones and load all
}
\examples{
\dontrun{
  pkg <- c("dplyr", "pacman")
  install_missing_load_all(pkg)
  }
}
