% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{lsos}
\alias{lsos}
\title{Memory check}
\source{
https://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session

Tony Breyal
}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{n}{how many objects to return}
}
\value{
printed description of objects in memory
}
\description{
Describe objects in memory
}
\examples{
lsos()

}
