% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{make_names}
\alias{make_names}
\title{Standardize variable names}
\usage{
make_names(names)
}
\arguments{
\item{names}{character vector to be coerced to syntactically valid names. This is coerced to character if necessary.}
}
\value{
A character vector of same length as \code{names} with each changed to a normalized variable name.
}
\description{
Normalize names to lowercase separated by underscores from character vector.
}
\examples{
n <- c('An', 'ex.amPLE', ' of me$$y', 'var__i@ble names  ')
make_names(n)
}
