% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_msg.R
\name{msg_guest_token}
\alias{msg_guest_token}
\title{Print GUEST_TOKEN to pass to auto_token argument of devtools::install_Github}
\usage{
msg_guest_token()
}
\value{
Message
}
\description{
Print GUEST_TOKEN to pass to auto_token argument of devtools::install_Github
}
\examples{
msg_guest_token()
}
