% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_msg.R
\name{msg_open}
\alias{msg_open}
\title{Print messages that commonly open documents.}
\usage{
msg_open()
}
\value{
Messages
}
\description{
Print messages that commonly open documents.
}
\examples{
msg_open()
}
