% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\source{
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\value{
arrangement of plots
}
\description{
Format multiple plots in a grid layout
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\examples{

library(ggplot2)
This example uses the ChickWeight dataset, which comes with ggplot2
First plot
p1 <- ggplot(ChickWeight, aes(x=Time, y=weight, colour=Diet, group=Chick)) +
geom_line() +
 ggtitle("Growth curve for individual chicks")

Second plot
p2 <- ggplot(ChickWeight, aes(x=Time, y=weight, colour=Diet)) +
 geom_point(alpha=.3) +
 geom_smooth(alpha=.2, size=1) +
 ggtitle("Fitted growth curve per diet")

multiplot(p1, p2)
}
