% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{named_list}
\alias{named_list}
\title{Named List}
\usage{
named_list(...)
}
\arguments{
\item{...}{objects to be passed into a list object}
}
\value{
A named list containing the input objects.
}
\description{
Set the names of a list object as the same as the object itself
}
\examples{
a <- TRUE
b <- 1:5
named_list(a, b)
#> $a
#> [1] TRUE
#>
#> $b
#> [1] 1 2 3 4 5

}
