% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{normVarNames}
\alias{normVarNames}
\title{Standardize variable names}
\usage{
normVarNames(df)
}
\arguments{
\item{df}{A dataframe to be coerced to have syntactically valid names.}
}
\value{
A dataframe with normalized variable names.
}
\description{
Normalize names to lowercase separated by underscores from character vector.
}
\examples{
d <- data.frame(list('An'=1, 'ex.amPLE'=2, ' of me$$y'=3, 'var__i@ble names  '=4))
normVarNames(d)
}
