% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open}
\alias{open}
\alias{open_snippets}
\alias{open_dir}
\alias{open_proj}
\alias{open_inst}
\alias{open_data}
\alias{open_data_raw}
\alias{open_docs}
\alias{open_man}
\alias{open_r}
\alias{open_tests}
\alias{open_testthat}
\title{Open files and directories of a project and package.}
\usage{
open_snippets()

open_dir(dir = NULL)

open_proj()

open_inst()

open_data()

open_data_raw()

open_docs()

open_man()

open_r()

open_tests()

open_testthat()
}
\description{
Open files and directories of a project and package.
}
