% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_rstudio_proj (Copia en conflicto de
%   Mauro TOSHIBA 2017-11-18).R
\name{open_snippets}
\alias{open_snippets}
\title{Open directory with snippets files.}
\usage{
open_snippets()
}
\description{
Open directory with snippets files.
}
