% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_arguments.R
\name{pass_arguments}
\alias{pass_arguments}
\title{Reformat arguments for funcitons, from arg1, arg2 to arg1 = arg1, arg2 = arg2}
\usage{
pass_arguments(string)
}
\arguments{
\item{string}{A string. e.g. c("arg1, arg2")}
}
\value{
A string. eg. arg1 = arg1, arg2 = arg2
}
\description{
This is useful when you extract a funciton and the arguments in the signature
need to be converted to arguments in a call to the function.
}
\examples{
pass_arguments(c("one, two"))
}
