% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_file.R
\name{prefix_file}
\alias{prefix_file}
\title{Add a prefix to a file name.}
\usage{
prefix_file(from, to = NULL, prefix)
}
\arguments{
\item{from}{Path where the files live.}

\item{to}{Optional. Path where the files will be copied to. If NULL (default)
files will be renamed in the directory \code{from}.}

\item{prefix}{A string. Prefix to add to the file names.}
}
\value{
The files with a prefix added to their names.
}
\description{
Add a prefix to a file name.
}
\examples{

prefix_file(from = "./file-from/", prefix = "my_prefix")
prefix_file(
  from = "./file-from/",
  to = "./file-to",
  prefix = "my_prefix"
)
}
