% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_msg.R
\name{print_zen}
\alias{print_zen}
\title{Print messages that commonly open documents.}
\usage{
print_zen()
}
\value{
Messages
}
\description{
Print messages that commonly open documents.
}
\examples{
pring_zen
}
