% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{remove_columns}
\alias{remove_columns}
\title{Remove columns based on a number of distinct values}
\usage{
remove_columns(df, .uniqueness = 1)
}
\arguments{
\item{df}{the input dataframe}

\item{.uniqueness}{the number of unique values the value should take on, omitting NAs.}
}
\value{
A dataframe with columns removed.
Note: using 0 will remove columns where all observations are NA.
}
\description{
Remove columns with less than a specified number of distinct values.
}
\details{
Functions to help with data cleaning
}
\examples{
df <- head(data.frame('x' = letters, 'y' = rep(0, length(letters)), z = rep(NA, length(letters))))
remove_columns(df)
remove_columns(df, 0)
}
