% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{removeColumns}
\alias{removeColumns}
\title{Remove columns based on a number of distinct values}
\usage{
removeColumns(df, .uniqueness = 1)
}
\arguments{
\item{df}{the input dataframe}

\item{.uniqueness}{the number of unique values the value should take on, omitting NAs.}
}
\value{
A dataframe with columns removed.
Note: using 0 will remove columns where all observations are NA.
}
\description{
Remove columns with less than a specified number of distinct values.
}
\examples{
\dontrun{
df <- head(data.frame('x' = letters, 'y' = rep(0, length(letters)), z = rep(NA, length(letters))))
removeColumns(df)
removeColumns(df, 0)
}
}
