% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusIS_1.R
\name{rename_chr}
\alias{rename_chr}
\title{rename chr name to 'chr1' mod.}
\usage{
rename_chr(chrs = c("NC_000012.12", "NC_000001.11"))
}
\arguments{
\item{chrs}{chrs}
}
\value{
chrs with new names formatted as 'chr1' to 'chrY'
}
\description{
rename chr name to 'chr1' mod.
}
