% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusIS_1.R
\name{seperate_gtf}
\alias{seperate_gtf}
\title{seperate the gtf for reads ext}
\usage{
seperate_gtf(
  input =
    "/home/SharedData/Ref_Data/GFF_Files_for_Gene_Annotation/hsa_GRCh38_latest_genomic.gff-sorted.bed"
)
}
\arguments{
\item{input}{sorted.bed}
}
\value{
bed files were writing into files.
}
\description{
seperate the gtf for reads ext
}
