% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusIS_1.R
\name{sort_bed_for_tss}
\alias{sort_bed_for_tss}
\title{Title sort the bed file for tss dist mapping.}
\usage{
sort_bed_for_tss(
  bed_path = "/home/SharedData/VIS/first_test/find-VIS-08-b-bed-from-sam.bed"
)
}
\arguments{
\item{bed_path}{path to bed file contains dist-to-tss}
}
\value{
dist-files
}
\description{
Title sort the bed file for tss dist mapping.
}
