% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{spreadColumn}
\alias{spreadColumn}
\title{Spread based on selected variables}
\usage{
spreadColumn(df, group_cols, spread_col, spread_val, fill_val = 0)
}
\arguments{
\item{df}{the input dataframe}

\item{group_cols}{the columns to group by}

\item{spread_col}{the column to spread by}

\item{spread_val}{the column values to spread}

\item{fill_val}{filling missing with what value}
}
\value{
dataframe that is spread based on values
}
\description{
Select and spread data
}
\examples{
df <- data.frame('id' = c(rep('x',2), rep('y',3),'z'), 'g' = letters[1:6], 'v' = 1:6)
spreadColumn(df, 'id','g','v')
}
