% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_lines.R
\name{start_lines_with}
\alias{start_lines_with}
\title{Add a string to the beginning of all lines in a file.}
\usage{
start_lines_with(file, start_with, save_as)
}
\arguments{
\item{file}{A path to a file.}

\item{start_with}{A string to start all lines with. When creating a function
from a long script, you may want to programmatically add two spaces to all
lines. So start_with "  ".}

\item{save_as}{If missing, the value  will be printed to the console.}
}
\description{
Add a string to the beginning of all lines in a file.
}
