% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\title{Produce a tag label of letters and numbers.}
\usage{
tag(n = 2)
}
\arguments{
\item{n}{Number of letters/numbers you want the tag to have.}
}
\value{
A string.
}
\description{
Produce a tag label of letters and numbers.
}
\examples{
tag()
tag(3)
}
