% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{title_names}
\alias{title_names}
\title{Make variable names into title case}
\usage{
title_names(df)
}
\arguments{
\item{df}{dataframe with names as character vector with underscore separation}
}
\value{
dataframe with names printed for reading
}
\description{
Replace underscores with spaces and make title case.
}
