% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_xy_value.R
\name{to_xy_value}
\alias{to_xy_value}
\title{Gather columns and rows of a matrix to variables x and y of a tibble.}
\usage{
to_xy_value(mat, grid)
}
\arguments{
\item{mat}{A matrix.}

\item{grid}{A named vector of length 2, with first element named x and second element named y.}
}
\value{
A tibble with variables \code{x}, \code{y}, \code{value}.
}
\description{
Gather columns and rows of a matrix to variables \code{x} and \code{y} of a tibble,
with values passed to variable \code{value}.
}
\examples{
\dontrun{
# Show that bci_elevation and bci_matrix contain the same information. They
# justdiffer in the class.
# library(tidyverse)
# library(bciex)
# 
# # Class is different
# class(bci_mat)
# class(bci_elevation)
# 
# # But contents are the same.
# # Tidy mat: each variable in a column; each observation in a row
# mat <- to_xy_value(bci_mat, grid = c(y = 5, x = 5))
# mat <- rename(mat, elev = value)
# # Arrange in the same way
# mat <- mat \%>\% arrange(x, y, elev)
# elev <- bci_elevation \%>\% arrange(x, y, elev)
# identical(mat, elev)
}
}
