% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusIS_1.R
\name{top10_exon_plot}
\alias{top10_exon_plot}
\title{get top 10 mapped gene.}
\usage{
top10_exon_plot(
  path_to_txt =
    "/home/SharedData/VIS/2022-07-second-sample/Rm_contam-06-gene-featurecounts-by-exon-cleaned.txt"
)
}
\arguments{
\item{path_to_txt}{input a path to the}
}
\value{
pic for top 10
}
\description{
get 10 mapped .
}
