% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusIS_1.R
\name{top10_site_plot}
\alias{top10_site_plot}
\title{Title get the percent plot for the top 10 sites.}
\usage{
top10_site_plot(
  cleaned_bedgraph =
    "/home/SharedData/VIS/2022-07-second-sample/Rm_contam-10-exon-renamed+sorted.bedgraph"
)
}
\arguments{
\item{cleaned_bedgraph}{cleaned bedgraph}
}
\value{
plot with the top10 VIS
}
\description{
Title get the percent plot for the top 10 sites.
}
