% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusIS_1.R
\name{tss_dist_plot}
\alias{tss_dist_plot}
\title{Title get tss dist plot from tss-dist.bed file}
\usage{
tss_dist_plot(
  tss_dist_bed =
    "/home/SharedData/VIS/2022-07-second-sample/demo/Res-11-gene-mapped-reads-dist_to_tss.bed",
  chr_col = 1,
  dist_cols = c(7, 11)
)
}
\arguments{
\item{tss_dist_bed}{bed contains tss_dist}

\item{chr_col}{num for chr column.}

\item{dist_cols}{list out the position from two files 列出连个文件中的位置所在的列，两个文件各列出1个碱基就够了。
The distance will be calculated again.}
}
\value{
pic by ggplot
}
\description{
Title get tss dist plot from tss-dist.bed file
}
