% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_rstudio_proj.R
\name{unlist_rstudio_proj}
\alias{unlist_rstudio_proj}
\title{Remove projects from RStudio's project list.}
\usage{
unlist_rstudio_proj(path_to_list = get_my_rstudio_proj_list(), rm_proj)
}
\arguments{
\item{path_to_list}{Path to the file where the projects are listed. In my
computer, they are given by \code{\link[=get_my_rstudio_proj_list]{get_my_rstudio_proj_list()}}.}

\item{rm_proj}{}
}
\value{
Invisible projects to keep. This function is called for its side
effect: to overwite the list of RStudio projects to exclude the ones the
user choosed to exclude.
}
\description{
Remove projects from RStudio's project list.
}
\examples{
\dontrun{
x <- unlist_rstudio_proj(get_my_rstudio_proj_list(), rm_proj = "todo|handy")
x
}
}
