% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadley.R
\name{unwhich}
\alias{unwhich}
\title{Convert integer to boolean representation.}
\usage{
unwhich(x, n)
}
\arguments{
\item{x}{A vector to which matches and missmatches to represent as TRUE and
FALSE.}

\item{n}{The length of the bolean representation.}
}
\value{
A boolean representation of length n.
}
\description{
Convert integer to boolean representation.
}
\examples{
x <- sample(10) < 4
which(x)
unwhich(which(x), 10)
}
