% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusIS_1.R
\name{write_bed}
\alias{write_bed}
\title{write dt to bed file}
\usage{
write_bed(dt, bed_file)
}
\arguments{
\item{dt}{data.table to be write}

\item{bed_file}{bed filename for output}
}
\description{
write dt to bed file
}
