% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fun_dependencies.R
\name{write_fun_dependencies}
\alias{write_fun_dependencies}
\title{Write a file with all dependencies of a user-defined function.}
\usage{
write_fun_dependencies(pkg, fun, file_nm)
}
\arguments{
\item{pkg}{String giving the package where \code{fun} lives, with the format
"package:PACKAGE", where lowercase is literal and caps is variable.}

\item{fun}{String giving the function which dependencies you want to write
to a file.}

\item{file_nm}{String giving the path to a file where to write the output.}
}
\value{
A file with the functions on which \code{fun} depends.
}
\description{
Write a file with all dependencies of a user-defined function.
}
\examples{
library(ctfs)
pkg <- "package:ctfs"
fun <- "growth"
file_nm <- "./dependencies_of_ctfs-growth.R"
write_fun_dependencies(pkg, fun, file_nm)
}
