% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{write_regression_to_excel}
\alias{write_regression_to_excel}
\title{Write a regression to Excel}
\usage{
write_regression_to_excel(reg_models, excel_file_name, sheets,
  title_names = TRUE)
}
\arguments{
\item{reg_models}{regression models, as a list}

\item{title_names}{whether to title format variable names}

\item{sheet_names}{formatted names of sheets}

\item{workbook_fname}{workbook file name to save to}
}
\value{
nothing
}
\description{
Writes a list of dataframe objects to an Excel workbook.
}
\examples{
df <- data.frame('a' = letters, 'b' = 1:length(letters), 'c' = rep(NA, length(letters)))
writeExcel(list(df),
           'DF',
           'wb.xlsx',
           title_names = TRUE)
}
