% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{writeExcel}
\alias{writeExcel}
\title{Write to Excel}
\usage{
writeExcel(dfs, sheet_names, workbook_fname, title_names = FALSE)
}
\arguments{
\item{dfs}{the input dataframes as a list}

\item{sheet_names}{formatted names of sheets}

\item{workbook_fname}{workbook file name to save to}

\item{title_names}{whether to title format variable names}
}
\value{
nothing
}
\description{
Writes a list of dataframe objects to an Excel workbook.
}
\examples{
df <- data.frame('a' = letters, 'b' = 1:length(letters), 'c' = rep(NA, length(letters)))
writeExcel(list(df),
           'DF',
           'wb.xlsx',
           title_names = TRUE)
}
