% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{edit_rstudio_recent_projects}
\alias{edit_rstudio_recent_projects}
\title{Open file to manually edit RStudio's list of Recent Projects}
\usage{
edit_rstudio_recent_projects()
}
\description{
Open file to manually edit RStudio's list of Recent Projects
}
