% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_rstudio_proj.R
\name{get_path_to_listed_rproj}
\alias{get_path_to_listed_rproj}
\title{Get the path to the file that stores RStudio's projects list.}
\usage{
get_path_to_listed_rproj(user = c("LeporeM", "dora"))
}
\arguments{
\item{user}{One of c("LeporeM", "dora").}
}
\value{
A path.
}
\description{
Get the path to the file that stores RStudio's projects list.
}
\examples{
get_path_to_listed_rproj()
}
