% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handy-package.R
\docType{package}
\name{handy-package}
\alias{handy}
\alias{handy-package}
\title{handy: Handy Tools for Common Problems}
\description{
Handy Tools for Common Problems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/handy}
  \item Report bugs at \url{https://github.com/maurolepore/handy}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com}

}
\keyword{internal}
