% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{make_names}
\alias{make_names}
\title{Standardize variable names}
\usage{
make_names(var_names)
}
\arguments{
\item{var_names}{vector of variable names}
}
\value{
cleaned up variable names
}
\description{
\code{make_names} returns a list that has been normalized to lowercase separated by underscores
}
