% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\title{Tag a line in a file with and optional message.}
\usage{
tag(msg = "")
}
\arguments{
\item{msg}{An optional message (default is an empty message).}
}
\value{
A closure (see \href{https://goo.gl/mHvapg}{Closures} in \href{http://adv-r.had.co.nz/}{AdvancedR} by Hadley Whickham).
}
\description{
Tags help track information because they can be searched in RStudio with the
shortcut \code{Ctrl+.} ("Go To File/Function...").
}
\examples{

# Tag a line in a file, so you can come back to that line later easily.
fig_key <- tag("Figure showhing key results.")
# <cool code>
# To locate the tag again, search it in RStudio with the shortcut Ctrl+.

# You may tag the next item in your "to do" list.
todo <- tag("Dear future me, continue working here.")
todo()

# Giving a message is optional.
todo <- tag()
todo()

# You may track tags in you manuscript, to link text and code.
}
