% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_rstudio_proj (Copia en conflicto de
%   Mauro TOSHIBA 2017-11-18).R, R/unlist_rstudio_proj.R
\name{unlist_rstudio_proj}
\alias{unlist_rstudio_proj}
\alias{unlist_rstudio_proj}
\title{Remove specific projects from RStudio's list under File > Recent Projects.}
\usage{
unlist_rstudio_proj(rm_proj, path = get_path_to_listed_rproj(user =
  c("LeporeM", "dora")))

unlist_rstudio_proj(rm_proj, path = get_path_to_listed_rproj(user =
  c("LeporeM", "dora")))
}
\arguments{
\item{rm_proj}{Pattern to match the project(s) to remove.}

\item{path}{Path to the file where RStudio stores the list of Recent
Projects.}

\item{rm_proj}{Pattern to match the project(s) to remove.}

\item{path}{Path to the file where RStudio stores the list of Recent
Projects.}
}
\value{
Invisibly returns the to Keep. The side effect is a list of Recent
RStudio projects thas lacks the ones the user chosed to remove.

Invisibly returns the to Keep. The side effect is a list of Recent
RStudio projects thas lacks the ones the user chosed to remove.
}
\description{
This function is called for its side effect: to remove one or more RStudio's
projects from the list of recent projects found in File > Recent Projects.

This function is called for its side effect: to remove one or more RStudio's
projects from the list of recent projects found in File > Recent Projects.
}
\examples{
\dontrun{
# Remove "project1" and "project3".
x <- unlist_rstudio_proj("project1|project3")
# Show the projects to keep in File > Recent Projects
x
}
\dontrun{
# Remove "project1" and "project3".
x <- unlist_rstudio_proj("project1|project3")
# Show the projects to keep in File > Recent Projects
x
}
}
