% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write_excel}
\alias{write_excel}
\title{Write to Excel}
\usage{
write_excel(dfs, workbook_fname, sheet_names, title_names = FALSE,
  add_row_border = FALSE)
}
\arguments{
\item{dfs}{the input dataframes as a list}

\item{workbook_fname}{workbook file name to save to}

\item{sheet_names}{(optional) formatted names of sheets, if not using object names as defaults}

\item{title_names}{(optional) whether to title format variable names}

\item{add_row_border}{(optional) whether to include an underline border under the names row}
}
\value{
nothing
}
\description{
Writes a list of dataframe objects to an Excel workbook.
}
\examples{
df <- data.frame('a' = letters, 'b' = 1:length(letters), 'c' = rep(NA, length(letters)))
write_excel(df,
            'wb.xlsx',
            title_names = TRUE)
}
