% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write_regression_to_excel}
\alias{write_regression_to_excel}
\title{Write a regression to Excel}
\usage{
write_regression_to_excel(reg_models, workbook_fname, sheet_names,
  title_names = TRUE)
}
\arguments{
\item{reg_models}{regression models, as a list}

\item{workbook_fname}{workbook file name to save to}

\item{sheet_names}{formatted names of sheets}

\item{title_names}{whether to title format variable names}
}
\value{
nothing
}
\description{
Writes a list of dataframe objects to an Excel workbook.
}
\examples{
# Basic Example
mod <- lm(mpg ~ cyl, mtcars)
\dontrun{
write_regression_to_excel(mod, 'simple_reg.xlsx')
}
# Writing a list of regressions
mod_hp <- lm(mpg ~ hp, mtcars)

mod_list <- list(mod_cyl = mod, mod_hp = mod_hp)
\dontrun{
write_regression_to_excel(mod_list, 'mpg_regs.xlsx')

write_regression_to_excel(mod_list, 'simple_reg.xlsx', c("Cylinder Model", "HP Model"))
}
}
